fixed selling issue

+++ carlcoin.js
@@ -1247,41 +1247,36 @@ client.on('message', message => {
 						let noUser = true;
 						for(let i=0;i<data.users.length;i++){
 							if(data.users[i].id == seller){
-								if(data.users[i].balance - price < 0){
-									message.channel.send('You dont have enough CC!');
-								}
-								else{
-									let noBuy = true;
-									for(let j=0;j<data.users.length;j++){
-										if(data.users[j].id == buyer){
-											let offerType = chop[chop.length-2]
-											//check if house
-											if(offerType == "house" || offerType == "apartment" || offerType == "skyscraper"){
-												if(data.users[j].balance - price < 0){
-													message.channel.send('Buyer doesnt have enough CC!');
-												}
-												else if(data.users[i][`${offerType}`] < 0 || isNaN(data.users[i][`${offerType}`])){
-													message.channel.send(`You dont have a ${offerType}`);
-												}
-												else{
-													let sellEnder = Date.now() + 60000;
-													let sellInfo = {"seller":`${seller}`,"sellerIndex":`${i}`,"buyer":`${buyer}`,"buyerIndex":`${j}`,"price":`${price}`,"sellEnder":`${sellEnder}`,"type":`${offerType}`};
-													let jsonBattle = JSON.stringify(sellInfo);
-													fs.writeFileSync(`/home/mattguy/carlcoin/cache/${buyer}houseSell`,jsonBattle);
-													message.channel.send(`${data.users[j].name}! Type !cc acceptPurchase to accept ${data.users[i].name}'s offer or type !cc denyPurchase to reject the offer. You have 1 minute to respond.`);
-													console.log(data.users[i].name + " trying to sell to " + data.users[j].name);
-												}
-											}
-											else{
+								let noBuy = true;
+								for(let j=0;j<data.users.length;j++){
+									if(data.users[j].id == buyer){
+										let offerType = chop[chop.length-2];
+										//check if house
+										if(offerType == "house" || offerType == "apartment" || offerType == "skyscraper"){
+											if(data.users[j].balance - price < 0){
 												message.channel.send('Buyer doesnt have enough CC!');
 											}
-											noBuy = false;
-											break;
+											else if(data.users[i][`${offerType}`] < 0 || isNaN(data.users[i][`${offerType}`])){
+												message.channel.send(`You dont have a ${offerType}`);
+											}
+											else{
+												let sellEnder = Date.now() + 60000;
+												let sellInfo = {"seller":`${seller}`,"sellerIndex":`${i}`,"buyer":`${buyer}`,"buyerIndex":`${j}`,"price":`${price}`,"sellEnder":`${sellEnder}`,"type":`${offerType}`};
+												let jsonBattle = JSON.stringify(sellInfo);
+												fs.writeFileSync(`/home/mattguy/carlcoin/cache/${buyer}houseSell`,jsonBattle);
+												message.channel.send(`${data.users[j].name}! Type !cc acceptPurchase to accept ${data.users[i].name}'s offer or type !cc denyPurchase to reject the offer. You have 1 minute to respond.`);
+												console.log(data.users[i].name + " trying to sell to " + data.users[j].name);
+											}
 										}
+										else{
+											message.channel.send('Buyer doesnt have enough CC!');
+										}
+										noBuy = false;
+										break;
 									}
-									if(noBuy){
-										message.channel.send('Buyer doesnt exist!');
-									}
+								}
+								if(noBuy){
+									message.channel.send('Buyer doesnt exist!');
 								}
 								noUser = false;
 								break;
@@ -1311,39 +1306,44 @@ client.on('message', message => {
 					message.channel.send('You have declined the offer');
 				}
 				else if(message.content.startsWith('!cc acceptPurchase')){
-					let database = fs.readFileSync('/home/mattguy/carlcoin/database.json');
-					let data = JSON.parse(database);
-					let price = parseInt(sellParse.price);
-					data.users[sellParse.sellerIndex].balance += price;
-					data.users[sellParse.buyerIndex].balance -= price;
-					if(sellParse.type == "house"){
-						data.users[sellParse.sellerIndex]["house"] -= 1;
-						data.users[sellParse.buyerIndex]["house"] += 1;
-						if(isNaN(data.users[sellParse.buyerIndex]["house"])){
-							data.users[sellParse.buyerIndex]["house"] = 1;
-						}
+					if(data.users[sellParse.buyerIndex].balance - sellParse.price < 0){
+						message.channel.send(`You don't have enough CC!`);
 					}
-					else if(sellParse.type == "apartment"){
-						data.users[sellParse.sellerIndex]["apartment"] -= 1;
-						data.users[sellParse.buyerIndex]["apartment"] += 1;
-						if(isNaN(data.users[sellParse.buyerIndex]["apartment"])){
-							data.users[sellParse.buyerIndex]["apartment"] = 1;
+					else{
+						let database = fs.readFileSync('/home/mattguy/carlcoin/database.json');
+						let data = JSON.parse(database);
+						let price = parseInt(sellParse.price);
+						data.users[sellParse.sellerIndex].balance += price;
+						data.users[sellParse.buyerIndex].balance -= price;
+						if(sellParse.type == "house"){
+							data.users[sellParse.sellerIndex]["house"] -= 1;
+							data.users[sellParse.buyerIndex]["house"] += 1;
+							if(isNaN(data.users[sellParse.buyerIndex]["house"])){
+								data.users[sellParse.buyerIndex]["house"] = 1;
+							}
 						}
-					}
-					else if(sellParse.type == "skyscraper"){
-						data.users[sellParse.sellerIndex]["skyscraper"] -= 1;
-						data.users[sellParse.buyerIndex]["skyscraper"] += 1;
-						if(isNaN(data.users[sellParse.buyerIndex]["skyscraper"])){
-							data.users[sellParse.buyerIndex]["skyscraper"] = 1;
+						else if(sellParse.type == "apartment"){
+							data.users[sellParse.sellerIndex]["apartment"] -= 1;
+							data.users[sellParse.buyerIndex]["apartment"] += 1;
+							if(isNaN(data.users[sellParse.buyerIndex]["apartment"])){
+								data.users[sellParse.buyerIndex]["apartment"] = 1;
+							}
 						}
+						else if(sellParse.type == "skyscraper"){
+							data.users[sellParse.sellerIndex]["skyscraper"] -= 1;
+							data.users[sellParse.buyerIndex]["skyscraper"] += 1;
+							if(isNaN(data.users[sellParse.buyerIndex]["skyscraper"])){
+								data.users[sellParse.buyerIndex]["skyscraper"] = 1;
+							}
+						}
+						data.users[sellParse.sellerIndex]["activity"] = Date.now();
+						data.users[sellParse.buyerIndex]["activity"] = Date.now();
+						let newData = JSON.stringify(data);
+						fs.writeFileSync('/home/mattguy/carlcoin/database.json',newData);
+						fs.unlinkSync(`/home/mattguy/carlcoin/cache/${personsId}houseSell`);
+						message.channel.send('You have accepted the offer!');
+						console.log(data.users[sellParse.sellerIndex].name + " has sold to " + data.users[sellParse.buyerIndex]);
 					}
-					data.users[sellParse.sellerIndex]["activity"] = Date.now();
-					data.users[sellParse.buyerIndex]["activity"] = Date.now();
-					let newData = JSON.stringify(data);
-					fs.writeFileSync('/home/mattguy/carlcoin/database.json',newData);
-					fs.unlinkSync(`/home/mattguy/carlcoin/cache/${personsId}houseSell`);
-					message.channel.send('You have accepted the offer!');
-					console.log(data.users[sellParse.sellerIndex].name + " has sold to " + data.users[sellParse.buyerIndex]);
 				}
 			}
 		}
